backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
groq==0.4.2
firebase-admin==6.2.0
python-dotenv==1.0.0
